DR16流管理发现严重问题

+++ justfw/interface/intf_dr16.h
@@ -59,7 +59,7 @@ typedef struct
     //    Key_t key[3];  // 改为位域后的键盘索引,空间减少8倍,速度增加16~倍
     //
     //    uint8_t key_count[3][16];
-    uint8_t sn;
+    uint16_t sn;
     uint32_t update_time;  // 更新时间，单位ms (HAL_GetTick())
 } RC_ctrl_t;
